arduino: change gesture type to uint8_t

+++ Arduino_sensing/Gestures.ino
@@ -32,7 +32,7 @@ void grabResponse(uint32_t startTime) {
   }
 }
 
-void gestureResponse(uint16_t gesture, uint32_t gestureStartTime) {
+void gestureResponse(uint8_t gesture, uint32_t gestureStartTime) {
   switch (gesture) {
     case 0:   noTone(TONE_PIN);                 break;
     case 1:   touchResponse(gestureStartTime);  break;
